(MINOR) Move exchange streaming client components to marketdata package (#901)
