+++ src/main/java/com/verlumen/tradestream/ingestion/BUILD
@@ -23,21 +23,6 @@ java_binary(
 )
 
 java_library(
-    name = "coinbase_streaming_client",
-    srcs = ["CoinbaseStreamingClient.java"],
-    deps = [
-        ":exchange_streaming_client",
-        "//protos:marketdata_java_proto",
-        "//src/main/java/com/verlumen/tradestream/instruments:currency_pair",
-        "//third_party:flogger",
-        "//third_party:gson",
-        "//third_party:guava",
-        "//third_party:guice",
-        "//third_party:protobuf_java_util",
-    ],
-)
-
-java_library(
     name = "coin_market_cap_config",
     srcs = ["CoinMarketCapConfig.java"],
 )
@@ -77,27 +62,6 @@ java_library(
         ":currency_pair_supply",
         "//src/main/java/com/verlumen/tradestream/http:http_client",
         "//third_party:gson",
-        "//third_party:guava",
-        "//third_party:guice",
-    ],
-)
-
-java_library(
-    name = "exchange_streaming_client",
-    srcs = ["ExchangeStreamingClient.java"],
-    deps = [
-        "//protos:marketdata_java_proto",
-        "//src/main/java/com/verlumen/tradestream/instruments:currency_pair",
-        "//third_party:guava",
-    ],
-)
-
-java_library(
-    name = "exchange_streaming_client_factory",
-    srcs = ["ExchangeStreamingClientFactory.java"],
-    deps = [
-        ":coinbase_streaming_client",
-        ":exchange_streaming_client",
         "//third_party:guava",
         "//third_party:guice",
     ],
@@ -153,8 +117,6 @@ java_library(
         ":coin_market_cap_config",
         ":currency_pair_supply",
         ":currency_pair_supply_provider",
-        ":exchange_streaming_client",
-        ":exchange_streaming_client_factory",
         ":ingestion_config",
         ":real_time_data_ingestion",
         ":real_time_data_ingestion_impl",
@@ -190,10 +152,10 @@ java_library(
     srcs = ["RealTimeDataIngestionImpl.java"],
     deps = [
         ":currency_pair_supply",
-        ":exchange_streaming_client",
         ":real_time_data_ingestion",
         "//protos:marketdata_java_proto",
         "//src/main/java/com/verlumen/tradestream/instruments:currency_pair",
+        "//src/main/java/com/verlumen/tradestream/marketdata:exchange_streaming_client",
         "//src/main/java/com/verlumen/tradestream/marketdata:trade_publisher",
         "//third_party:flogger",
         "//third_party:guava",
