+++ src/main/java/com/verlumen/tradestream/ingestion/IngestionModule.java
@@ -19,24 +19,17 @@ abstract class IngestionModule extends AbstractModule {
   @Override
   protected void configure() {
     bind(CurrencyPairSupply.class).toProvider(CurrencyPairSupplyProvider.class);
-    bind(ExchangeStreamingClient.Factory.class).to(ExchangeStreamingClientFactory.class);
     bind(RealTimeDataIngestion.class).to(RealTimeDataIngestionImpl.class);
 
     install(HttpModule.create());
     install(KafkaModule.create(ingestionConfig().kafkaBootstrapServers()));
-    install(MarketDataModule.create(ingestionConfig().tradeTopic()));
+    install(MarketDataModule.create(ingestionConfig().exchangeName(), ingestionConfig().tradeTopic()));
   }
 
   @Provides
   CoinMarketCapConfig provideCoinMarketCapConfig() {
     return CoinMarketCapConfig.create(
         ingestionConfig().topCryptocurrencyCount(), ingestionConfig().coinMarketCapApiKey());
-  }
-
-  @Provides
-  ExchangeStreamingClient provideExchangeStreamingClient(
-      ExchangeStreamingClient.Factory exchangeStreamingClientFactory) {
-    return exchangeStreamingClientFactory.create(ingestionConfig().exchangeName());
   }
 
   @Provides
