+++ src/main/java/com/verlumen/tradestream/marketdata/BUILD
@@ -18,6 +18,22 @@ java_library(
     ],
 )
 
+
+java_library(
+    name = "coinbase_streaming_client",
+    srcs = ["CoinbaseStreamingClient.java"],
+    deps = [
+        ":exchange_streaming_client",
+        "//protos:marketdata_java_proto",
+        "//src/main/java/com/verlumen/tradestream/instruments:currency_pair",
+        "//third_party:flogger",
+        "//third_party:gson",
+        "//third_party:guava",
+        "//third_party:guice",
+        "//third_party:protobuf_java_util",
+    ],
+)
+
 java_library(
     name = "default_trade_generator",
     srcs = ["DefaultTradeGenerator.java"],
@@ -28,6 +44,27 @@ java_library(
       "//third_party:joda_time",
       "//third_party:protobuf_java",
       "//third_party:protobuf_java_util",
+    ],
+)
+
+java_library(
+    name = "exchange_streaming_client",
+    srcs = ["ExchangeStreamingClient.java"],
+    deps = [
+        "//protos:marketdata_java_proto",
+        "//src/main/java/com/verlumen/tradestream/instruments:currency_pair",
+        "//third_party:guava",
+    ],
+)
+
+java_library(
+    name = "exchange_streaming_client_factory",
+    srcs = ["ExchangeStreamingClientFactory.java"],
+    deps = [
+        ":coinbase_streaming_client",
+        ":exchange_streaming_client",
+        "//third_party:guava",
+        "//third_party:guice",
     ],
 )
 
@@ -52,6 +89,8 @@ java_library(
   name = "market_data_module",
   srcs = ["MarketDataModule.java"],
   deps = [
+    ":exchange_streaming_client",
+    ":exchange_streaming_client_factory",
     ":market_data_config",
     ":trade_publisher",
     ":trade_publisher_impl",
