+++ src/main/java/com/verlumen/tradestream/marketdata/ExchangeStreamingClient.java
@@ -1,11 +1,9 @@
-package com.verlumen.tradestream.ingestion;
+package com.verlumen.tradestream.marketdata;
 
 import static com.google.common.collect.ImmutableList.toImmutableList;
 
 import com.google.common.collect.ImmutableList;
 import com.verlumen.tradestream.instruments.CurrencyPair;
-import com.verlumen.tradestream.marketdata.Trade;
-
 import java.util.function.Consumer;
 
 /**
@@ -13,7 +11,7 @@ import java.util.function.Consumer;
  * This allows different exchanges to provide their own implementations
  * while maintaining a consistent interface for the data ingestion system.
  */
-interface ExchangeStreamingClient {
+public interface ExchangeStreamingClient {
     /**
      * Starts streaming market data for the given currency pairs.
      *
