+++ src/main/java/com/verlumen/tradestream/marketdata/MarketDataConfig.java
@@ -1,7 +1,7 @@
 package com.verlumen.tradestream.marketdata;
 
-public record MarketDataConfig(String tradeTopic) {
-  public static MarketDataConfig create(String tradeTopic) {
-    return new MarketDataConfig(tradeTopic);
+public record MarketDataConfig(String exchangeName, String tradeTopic) {
+  public static MarketDataConfig create(String exchangeName, String tradeTopic) {
+    return new MarketDataConfig(exchangeName, tradeTopic);
   }
 }
